chore: clean up nixos modules and fix bad conditional in zshrc

+++ modules/editor/nvim/lua/ui/terminal/config.lua
@@ -8,7 +8,7 @@ return {
         col = 0.25,
         width = 0.5,
         height = 0.4,
-        border = "double",
+        border = "single",
       },
       horizontal = { location = "rightbelow", split_ratio = 0.3 },
       vertical = { location = "rightbelow", split_ratio = 0.5 },
@@ -23,4 +23,3 @@ return {
     auto_insert = true,
   },
 }
-
